requirements.txt
asgiref==3.7.2
certifi==2023.11.17
contourpy==1.2.0
cycler==0.12.1
Django==4.2.7
django-ratelimit2==0.4.2
djangorestframework==3.14.0
fonttools==4.46.0
idna==3.6
kiwisolver==1.4.5
Markdown==3.5.2
matplotlib==3.8.2
mysqlclient==2.2.0
numpy==1.26.2
packaging==23.2
Pillow==10.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
urllib3==2.1.0

pip~=23.3
wheel~=0.41.2
setuptools~=68.0.0

django-widget-tweaks==1.5.0

django-widget-tweaks==1.5.0

django-storages[google]
gcsUtils
